Rename AlertMessageManager and move to separate file
